Add snippet validation test to test_utils.py
